Move the joint_state_estimator into its own folder/
Since the JointStateEstimator stuff doesn't in the LFC, we consider it as an
additional/external tool, yet no nodes nor plugins are created from it...

It should either be removed, moved into its own repo or be used in order to
define an executable (e.g. ros node, ros control plugin, ...).
In order to anticipate those, we isolated it to ease futur actions on it.